Wrap the script in a python package
